+++ java11/src/main/java/feign/http2client/Http2Client.java
@@ -36,7 +36,7 @@ import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class Http2Client implements Client {
+public class Http2Client extends AbstractHttpClient implements Client {
 
   private final HttpClient client;
 
@@ -53,7 +53,14 @@ public class Http2Client implements Client {
 
   @Override
   public Response execute(Request request, Options options) throws IOException {
-    final HttpRequest httpRequest = newRequestBuilder(request, options).build();
+    final HttpRequest httpRequest;
+    try {
+      httpRequest = newRequestBuilder(request, options)
+              .version(Version.HTTP_2)
+              .build();
+    } catch (URISyntaxException e) {
+      throw new IOException("Invalid uri " + request.url(), e);
+    }
 
     HttpResponse<byte[]> httpResponse;
     try {
@@ -63,59 +70,7 @@ public class Http2Client implements Client {
       throw new IOException("Invalid uri " + request.url(), e);
     }
 
-    final OptionalLong length = httpResponse.headers().firstValueAsLong("Content-Length");
-
-    final Response response = Response.builder()
-        .body(new ByteArrayInputStream(httpResponse.body()),
-            length.isPresent() ? (int) length.getAsLong() : null)
-        .reason(httpResponse.headers().firstValue("Reason-Phrase").orElse("OK"))
-        .request(request)
-        .status(httpResponse.statusCode())
-        .headers(castMapCollectType(httpResponse.headers().map()))
-        .build();
-    return response;
-  }
-
-  private Builder newRequestBuilder(Request request, Options options) throws IOException {
-    URI uri;
-    try {
-      uri = new URI(request.url());
-    } catch (final URISyntaxException e) {
-      throw new IOException("Invalid uri " + request.url(), e);
-    }
-
-    final BodyPublisher body;
-    final byte[] data = request.body();
-    if (data == null) {
-      body = BodyPublishers.noBody();
-    } else {
-      body = BodyPublishers.ofByteArray(data);
-    }
-
-    final Builder requestBuilder = HttpRequest.newBuilder()
-        .uri(uri)
-        .timeout(Duration.ofMillis(options.readTimeoutMillis()))
-        .version(Version.HTTP_2);
-
-    final Map<String, Collection<String>> headers = filterRestrictedHeaders(request.headers());
-    if (!headers.isEmpty()) {
-      requestBuilder.headers(asString(headers));
-    }
-
-    switch (request.httpMethod()) {
-      case GET:
-        return requestBuilder.GET();
-      case POST:
-        return requestBuilder.POST(body);
-      case PUT:
-        return requestBuilder.PUT(body);
-      case DELETE:
-        return requestBuilder.DELETE();
-      default:
-        // fall back scenario, http implementations may restrict some methods
-        return requestBuilder.method(request.httpMethod().toString(), body);
-    }
-
+    return toFeignResponse(request, httpResponse);
   }
 
   /**
@@ -133,7 +88,8 @@ public class Http2Client implements Client {
     DISALLOWED_HEADERS_SET = Collections.unmodifiableSet(treeSet);
   }
 
-  private Map<String, Collection<String>> filterRestrictedHeaders(Map<String, Collection<String>> headers) {
+  @Override
+  protected Map<String, Collection<String>> filterRestrictedHeaders(Map<String, Collection<String>> headers) {
     final Map<String, Collection<String>> filteredHeaders = headers.keySet()
         .stream()
         .filter(headerName -> !DISALLOWED_HEADERS_SET.contains(headerName))
@@ -145,20 +101,4 @@ public class Http2Client implements Client {
 
     return filteredHeaders;
   }
-
-  private Map<String, Collection<String>> castMapCollectType(Map<String, List<String>> map) {
-    final Map<String, Collection<String>> result = new HashMap<>();
-    map.forEach((key, value) -> result.put(key, new HashSet<>(value)));
-    return result;
-  }
-
-  private String[] asString(Map<String, Collection<String>> headers) {
-    return headers.entrySet().stream()
-        .flatMap(entry -> entry.getValue()
-            .stream()
-            .map(value -> Arrays.asList(entry.getKey(), value))
-            .flatMap(List::stream))
-        .toArray(String[]::new);
-  }
-
 }
